Rewrite client interface for login server (part 7)
Added private interface in lclif.p.h
