+++ src/login/HPMlogin.c
@@ -26,6 +26,7 @@
 
 #include "login/account.h"
 #include "login/lclif.h"
+#include "login/lclif.p.h"
 #include "login/login.h"
 #include "common/HPMi.h"
 #include "common/conf.h"
