+++ src/login/Makefile.in
@@ -43,7 +43,7 @@ MT19937AR_H = $(MT19937AR_D)/mt19937ar.h
 LOGIN_C = account_sql.c HPMlogin.c ipban_sql.c lclif.c login.c loginlog_sql.c
 LOGIN_OBJ = $(addprefix obj_sql/, $(patsubst %.c,%.o,$(LOGIN_C)))
 LOGIN_H = login.h account.h HPMlogin.h ipban.h lclif.h loginlog.h
-LOGIN_PH =
+LOGIN_PH = lclif.p.h
 
 HAVE_MYSQL=@HAVE_MYSQL@
 ifeq ($(HAVE_MYSQL),yes)
