+++ src/plugins/HPMHooking.c
@@ -30,6 +30,7 @@
 #define HPM_POINTS_INCLUDE "HPMHooking/HPMHooking_login.HookingPoints.inc"
 #define HPM_SOURCES_INCLUDE "HPMHooking/HPMHooking_login.sources.inc"
 #include "login/lclif.h"
+#include "login/lclif.p.h"
 #include "login/login.h"
 #elif defined (HPMHOOKING_CHAR)
 #define HPM_SERVER_TYPE SERVER_TYPE_CHAR
